Move conversion to confirm screen to handle success message
