+++ mobile-app/app/components/InfoText.tsx
@@ -8,14 +8,14 @@ interface InfoTextProp extends ThemedProps, TextProps {
   type?: InfoTextType
 }
 
-export type InfoTextType = 'warning' | 'error'
+export type InfoTextType = 'warning' | 'error' | 'success'
 
 export function InfoText (props: InfoTextProp): JSX.Element {
   const {
     type = 'warning',
     style,
-    light = tailwind({ 'bg-warning-50 border-warning-200': type === 'warning', 'bg-error-50 border-error-200': type === 'error' }),
-    dark = tailwind({ 'bg-darkwarning-50 border-darkwarning-200': type === 'warning', 'bg-darkerror-50 border-darkerror-200': type === 'error' }),
+    light = tailwind({ 'bg-warning-50 border-warning-200': type === 'warning', 'bg-error-50 border-error-200': type === 'error', 'bg-success-50 border-success-200': type === 'success' }),
+    dark = tailwind({ 'bg-darkwarning-50 border-darkwarning-200': type === 'warning', 'bg-darkerror-50 border-darkerror-200': type === 'error', 'bg-success-50 border-success-200': type === 'success' }),
     ...otherProps
   } = props
 
@@ -30,10 +30,10 @@ export function InfoText (props: InfoTextProp): JSX.Element {
     >
       <ThemedIcon
         iconType='MaterialIcons'
-        name={type === 'warning' ? 'info' : 'warning'}
+        name={type === 'success' ? 'check-circle-outline' : type === 'warning' ? 'info' : 'warning'}
         size={14}
-        light={tailwind({ 'text-warning-500': type === 'warning', 'text-error-500': type === 'error' })}
-        dark={tailwind({ 'text-darkwarning-500': type === 'warning', 'text-darkerror-500': type === 'error' })}
+        light={tailwind({ 'text-warning-500': type === 'warning', 'text-error-500': type === 'error', 'text-success-500': type === 'success' })}
+        dark={tailwind({ 'text-darkwarning-500': type === 'warning', 'text-darkerror-500': type === 'error', 'text-darksuccess-500': type === 'success' })}
       />
       <ThemedText
         style={tailwind('text-xs pl-2 font-medium flex-1')}
